requirements.txt
folium==0.16.0
haversine==2.8.1
numpy==1.24.3
pandas==2.0.3
Pillow==9.4.0
plotly==5.9.0
streamlit==1.35.0
streamlit_folium==0.20.0
matplotlib==3.5.3
matplotlib-inline==0.1.6